Move slice tests to a new file
